+++ packages/app-btc/src/index.ts
@@ -1,4 +1,5 @@
 export * from './app';
 export * from './proto/generated/types';
 export * from './operations/types';
-export { updateLogger } from './utils';
+export { updateLogger, setBitcoinJSLib } from './utils';
+export type { bitcoinJsLibType } from './utils';