+++ packages/app-btc/src/operations/getPublicKey/types.ts
@@ -8,3 +8,8 @@ export type GetPublicKeyEventHandler = (event: GetPublicKeyStatus) => void;
 export interface IGetPublicKeyParams extends IGetPublicKeyIntiateRequest {
   onEvent?: GetPublicKeyEventHandler;
 }
+
+export interface IGetPublicKeyResult {
+  publicKey: Uint8Array;
+  address: string;
+}